refactor: moved initRouterStore to useRouterStore file
