+++ packages/tuono/src/router/components/RouterProvider.tsx
@@ -1,8 +1,9 @@
 import { getRouterContext } from './RouterContext'
 import { Matches } from './Matches'
-import { useRouterStore } from '../hooks/useRouterStore'
 import { useListenBrowserUrlUpdates } from '../hooks/useListenBrowserUrlUpdates.tsx'
-import React, { useEffect, useLayoutEffect, type ReactNode } from 'react'
+import React, { type ReactNode } from 'react'
+import { initRouterStore } from '../hooks/useRouterStore'
+import type { ServerProps } from '../types'
 
 type Router = any
 
@@ -14,11 +15,6 @@ interface RouterContextProviderProps {
 interface RouterProviderProps {
   router: Router
   serverProps?: ServerProps
-}
-
-interface ServerProps {
-  router: Location
-  props: any
 }
 
 function RouterContextProvider({
@@ -46,30 +42,6 @@ function RouterContextProvider({
       <routerContext.Provider value={router}>{children}</routerContext.Provider>
     </React.Suspense>
   )
-}
-
-const initRouterStore = (props?: ServerProps): void => {
-  const updateLocation = useRouterStore((st) => st.updateLocation)
-
-  if (typeof window === 'undefined') {
-    updateLocation({
-      pathname: props?.router.pathname || '',
-      hash: '',
-      href: '',
-      searchStr: '',
-    })
-  }
-
-  useLayoutEffect(() => {
-    const { pathname, hash, href, search } = window.location
-    updateLocation({
-      pathname,
-      hash,
-      href,
-      searchStr: search,
-      search: new URLSearchParams(search),
-    })
-  }, [])
 }
 
 export function RouterProvider({
